Add files via upload
This is a probabilistic volume calculation sheet for oil and gas volumes initially in place. It is performed through the Monte Carlo simulation method which is widely used in the industry to calculate possible volumes in a reservoir during exploration.  I developed this tool because I didn't want to buy expensive software to calculate these volumes while we can simply write a simple code to do it for us while at the same time controlling what is in the code and using it for free.

Feel free to send me comments or improvement suggestions.